requirements.txt
# Core Computer Vision
ultralytics>=8.0.0
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
scipy>=1.11.0

# FastAPI Backend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Configuration & Utilities
pyyaml>=6.0
python-dotenv>=1.0.0

# Optional: GPU acceleration
# torch>=2.0.0
# torchvision>=0.15.0
